Add additional check for sorting local changes
It was possible that multiple local changes are created with the same "publishDate". We would then not be sure if we apply change 1 before change 3, because they have the same "publishDate".
Now I sort by "publishDate" ascending, and then by the change id ascending.
This should make sure that we always apply all changes in the correct order.